fix: Split utility functions into multiple files to improve maintainability
